+++ src/main/java/dev/muffin/rpgcore/rpg/rpgutils/ArchetypeConstants.java
@@ -4,7 +4,7 @@ public class ArchetypeConstants {
 
     // Warrior
     public static final double WARRIOR_BASE_HP = 250;
-    public static final double WARRIOR_BASE_MANA = 70;
+    public static final int WARRIOR_BASE_MANA = 70;
     public static final double WARRIOR_BASE_MANA_REGEN = 5;
 
 }
