+++ src/main/java/dev/muffin/rpgcore/rpg/rpgutils/RPGClassStats.java
@@ -3,10 +3,10 @@ package dev.muffin.rpgcore.rpg.rpgutils;
 public class RPGClassStats {
 
     public final double hp;
-    public final double mana;
-    public final double manaRegen;
+    public final int mana;
+    public final int manaRegen;
 
-    public RPGClassStats(double hp, double mana, double manaRegen) {
+    public RPGClassStats(double hp, int mana, int manaRegen) {
         this.hp = hp;
         this.mana = mana;
         this.manaRegen = manaRegen;
